Fixed variable reference in database cleanup
Sometimes a text channel will get deleted (or become inaccesible to the
bot), and a subscription entry in the database will need to get deleted
automatically. There was an undeclared variable reference in the section
of code that pushes the subscriptions, specifically in the part that
wipes a subscription entry if a channel becomes deleted/inaccesible. I
did not catch this error originally because the conditions nessesary to
cause it, namely the deletion of channels that did not have their
subscriptions removed manually, did not occur until I deleted one of my
test servers.