Fixed glitches with Catholic lectionary parsing
Originally, if a reference in the Responsorial Psalm section did not
start with "Ps", this was inserted automatically. However, this broke down
if a verse reference for another book was present explicitly, say
something from Isaiah; it would append the Psalm book reference
mistakenly. This was corrected by checking if the reference starts with
a number, which would indicate whether a book reference was actually
missing. Secondly, sometimes a reference would begin with "Cf". A
parsing case was added to make sure this doesn't get mistakenly included
in the Bible reference link.